Move error to its own file
